More Control Clarity
Make changes to states or edges by right-clicking them.

Multi-selections can now all be deleted at once.

If holding the multi-select button (shift by default), a newly inserted state is added to the selection.

When inserting a new state it can be dragged so long as the mouse button is held.